move `datastore_test` to conftest.py
