+++ src/middlewared/middlewared/pytest/unit/plugins/test_datastore.py
@@ -1,11 +1,9 @@
-from contextlib import asynccontextmanager
 import datetime
 from unittest.mock import ANY, patch
 
 import pytest
 import sqlalchemy as sa
 from sqlalchemy.exc import IntegrityError
-from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import relationship
 
 import middlewared.plugins.datastore  # noqa
@@ -13,49 +11,8 @@ import middlewared.plugins.datastore.connection  # noqa
 import middlewared.plugins.datastore.schema  # noqa
 import middlewared.plugins.datastore.util  # noqa
 from middlewared.plugins.datastore.write import NoRowsWereUpdatedException
-from middlewared.pytest.unit.helpers import load_compound_service
-from middlewared.pytest.unit.middleware import Middleware
+from middlewared.pytest.unit.plugins.conftest import Model, datastore_test
 from middlewared.sqlalchemy import JSON, EncryptedText, Time
-
-DatastoreService = load_compound_service("datastore")
-
-Model = declarative_base()
-
-
-@asynccontextmanager
-async def datastore_test(mocked_calls=None):
-    mocked_calls = mocked_calls or {}
-    m = Middleware()
-    with patch("middlewared.plugins.datastore.connection.FREENAS_DATABASE", ":memory:"):
-        with patch("middlewared.plugins.datastore.schema.Model", Model):
-            with patch("middlewared.plugins.datastore.util.Model", Model):
-                ds = DatastoreService(m)
-                ds.setup()
-
-                for part in ds.parts:
-                    if hasattr(part, "connection"):
-                        Model.metadata.create_all(bind=part.connection)
-                        break
-                else:
-                    raise RuntimeError("Could not find part that provides connection")
-
-                m["datastore.execute"] = ds.execute
-                m["datastore.execute_write"] = ds.execute_write
-                m["datastore.fetchall"] = ds.fetchall
-
-                m["datastore.query"] = ds.query
-                m["datastore.send_insert_events"] = ds.send_insert_events
-                m["datastore.send_update_events"] = ds.send_update_events
-                m["datastore.send_delete_events"] = ds.send_delete_events
-
-                m["datastore.insert"] = ds.insert
-                m["datastore.update"] = ds.update
-                m["datastore.delete"] = ds.delete
-
-                for call_name, call_func in mocked_calls.items():
-                    m[call_name] = call_func
-
-                yield ds
 
 
 class UserModel(Model):
